fix: reuse Twitch callback for linked account auth
